requirements.txt
# 필수 패키지
Pillow>=10.0.0
numpy>=1.24.0

# 선택적 패키지 (얼굴 인식 기능 사용 시 필요)
opencv-python>=4.8.0

# 선택적 패키지 (얼굴 랜드마크 기능 사용 시 필요)
mediapipe>=0.10.0

# 선택적 패키지 (Delaunay Triangulation 기반 변형 기능 사용 시 필요)
scipy>=1.10.0